+++ hypothesis-python/src/hypothesis/_settings.py
@@ -561,6 +561,23 @@ class HealthCheck(Enum, metaclass=HealthCheckMeta):
     subclasses, or to refactor so that :func:`@given <hypothesis.given>` is
     specified on leaf subclasses."""
 
+    nested_given = 11
+    """Checks if :func:`@given <hypothesis.given>` is used inside another
+    :func:`@given <hypothesis.given>`. This results in quadratic generation and
+    shrinking behavior, and can usually be expressed more cleanly by using
+    :func:`~hypothesis.strategies.data` to replace the inner
+    :func:`@given <hypothesis.given>`.
+
+    Nesting @given can be appropriate if you set appropriate limits for the
+    quadratic behavior and cannot easily reexpress the inner function with
+    :func:`~hypothesis.strategies.data`. To suppress this health check, set
+    ``suppress_health_check=[HealthCheck.nested_given]`` on the outer
+    :func:`@given <hypothesis.given>`. Setting it on the inner
+    :func:`@given <hypothesis.given>` has no effect. If you have more than one
+    level of nesting, add a suppression for this health check to every
+    :func:`@given <hypothesis.given>` except the innermost one.
+    """
+
 
 @unique
 class Verbosity(IntEnum):