+++ hypothesis-python/src/hypothesis/strategies/_internal/recursive.py
@@ -47,10 +47,10 @@ class LimitedStrategy(SearchStrategy):
     def currently_capped(self, value):
         self._threadlocal.currently_capped = value
 
-    def __repr__(self):
+    def __repr__(self) -> str:
         return f"LimitedStrategy({self.base_strategy!r})"
 
-    def do_validate(self):
+    def do_validate(self) -> None:
         self.base_strategy.validate()
 
     def do_draw(self, data):
@@ -83,7 +83,7 @@ class RecursiveStrategy(SearchStrategy):
             strategies.append(extend(OneOfStrategy(tuple(strategies))))
         self.strategy = OneOfStrategy(strategies)
 
-    def __repr__(self):
+    def __repr__(self) -> str:
         if not hasattr(self, "_cached_repr"):
             self._cached_repr = "recursive(%r, %s, max_leaves=%d)" % (
                 self.base,
@@ -92,7 +92,7 @@ class RecursiveStrategy(SearchStrategy):
             )
         return self._cached_repr
 
-    def do_validate(self):
+    def do_validate(self) -> None:
         check_strategy(self.base, "base")
         extended = self.extend(self.limited_base)
         check_strategy(extended, f"extend({self.limited_base!r})")