+++ hypothesis-python/tests/nocover/test_randomization.py
@@ -12,7 +12,14 @@ from random import Random
 
 import pytest
 
-from hypothesis import Verbosity, core, given, settings, strategies as st
+from hypothesis import (
+    HealthCheck,
+    Verbosity,
+    core,
+    given,
+    settings,
+    strategies as st,
+)
 
 from tests.common.utils import Why, no_shrink, xfail_on_crosshair
 
@@ -46,7 +53,12 @@ def test_seeds_off_internal_random():
 @xfail_on_crosshair(Why.nested_given)
 def test_nesting_with_control_passes_health_check():
     @given(st.integers(0, 100), st.random_module())
-    @settings(max_examples=5, database=None, deadline=None)
+    @settings(
+        max_examples=5,
+        database=None,
+        deadline=None,
+        suppress_health_check=[HealthCheck.nested_given],
+    )
     def test_blah(x, rnd):
         @given(st.integers())
         @settings(